item fait + ressoureces

+++ app/src/main/java/com/example/tp1_consortium/domain/models/Delivery.kt
@@ -4,11 +4,11 @@ import androidx.room.Entity
 import androidx.room.PrimaryKey
 @Entity(tableName = "deliveries")
 data class Delivery (
-                var jasmalt: Double = 0.0,
-                var kreotrium: Double = 0.0,
-                var xuskian: Double = 0.0,
-                var yefrium:Double=0.0,
-                var zuscum:Double=0.0 ) {
+                var jasmalt: Float = 0.0f,
+                var kreotrium: Float = 0.0f,
+                var xuskian: Float = 0.0f,
+                var yefrium:Float=0.0f,
+                var zuscum:Float=0.0f ) {
     @PrimaryKey(autoGenerate = true)
     var idDelivery : Int = 0
 }
